Structure changes
Due to difficulties in testing the clients changes have been made:
1. Creation of interface Client with a general implementation of it(also
to avoid repeating code).
2. Two existing client are extending the general client. some common
code moved to the parent.

Additional of explanation added.

Removal of a flag in QuizImpl.
This flag was made to avoid a situation of client removing a quiz that
is played at the same moment.
The quiz can still be played but the result would not set as a record.